% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_manifest.R
\name{check_arg_type}
\alias{check_arg_type}
\title{check_arg_type}
\usage{
check_arg_type(arg)
}
\arguments{
\item{arg}{object to check. Lists will be checked recursively, and must be
named.}
}
\value{
the same object, unchanged. Function will throw an error if objects are not simple
}
\description{
Check that arguments are nicely coercible to JSON. Primarily a check that
lists are composed of simple types (other lists, characters,
numeric/integers, or logicals). Called for side effect of \code{stop} if not.
}
