% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_checks.R
\name{check_dir_writable}
\alias{check_dir_writable}
\title{check_dir_writable}
\usage{
check_dir_writable(dir)
}
\arguments{
\item{dir}{directory to check writiblity}
}
\value{
logical, can a file be written to the directory?
}
\description{
Check if a directory is writable. Check is performed by attempting to write
a 0-byte file (\code{.test}) to the directory (and removing afterwards). Emits a
warning on windows, since file permissions on Windows systems are difficult
to test, but \emph{should} be accurate.
}
