% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_checks.R
\name{check_file}
\alias{check_file}
\title{check_file}
\usage{
check_file(filepath)
}
\arguments{
\item{filepath}{path to file}
}
\value{
logical(1L), is the file valid?
}
\description{
check a filepath for existence, non-zero size, and non-directory status.
}
