% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_checks.R
\name{check_io}
\alias{check_io}
\title{check_io}
\usage{
check_io(input_files = NULL, output_dirs = NULL)
}
\arguments{
\item{input_files}{list or vector of file paths to check with
\code{\link{check_file}}}

\item{output_dirs}{list or vector of directories to check with
\code{\link{check_dir_writable}}}
}
\value{
logical(1L), are all files and directories valid? Note that function
will throw an error in situations where this is not TRUE.
}
\description{
check \code{input_files} using \code{\link{check_file}} and \code{output_dirs} using
\code{\link{check_dir_writable}}.
}
