% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{get_git_info}
\alias{get_git_info}
\title{get_git_info}
\usage{
get_git_info(repo)
}
\arguments{
\item{repo}{(Local directory) path to a git repository.}
}
\value{
list of git information. Keys include:
\itemize{
\item repo: path to the repository.
\item is_git: logical indicating if the path is in a git repository.
\item commit: latest commit hash.
\item clean: logical indicating if the repository is clean.
\item branch: list of branch information. See \code{git_branch_info}.
\item changed_files: list of changed files. See \code{git_changed_files}.
\item tags: list of tags. See \code{git_tag_info}.
}
}
\description{
Get git information for a repository.
}
