% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_info.R
\name{get_individual_package_info}
\alias{get_individual_package_info}
\title{Get package information for a package}
\usage{
get_individual_package_info(packagename)
}
\arguments{
\item{packagename}{Singular charater string of package name}
}
\value{
nested list of file details, length 11, with keys:
\itemize{
\item \code{package}: The name of the package
\item \code{version}: The version of the package
\item \code{loaded_with_pkgload}: Is this package loaded with \code{pkgload}? (logical).
Useful for identifying local development versions
\item \code{library}: The path of the library the package is installed in
\item \code{library_index}: The index of the library in the \code{.libPaths()} vector
\item \code{repository}: The repository the package was pulled from
\item \code{platform}: The platform the package was built for
\item \code{built}: Information about package build (relevant for binary packages)
\item \code{remotetype}: The type of remote repository the package was pulled from
\item \code{remotepkgref}: The reference used by \code{pak} to install the package
\item \code{remoteref}: The reference of the package when it was pulled from REPO
\item \code{remotesha}: the SHA-1 hash of the reference (if applicable)
}
}
\description{
This function takes a single package name and returns a list of package
details, suitable for inclusion in manifest export.
}
