% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_environment.R
\name{get_manifest_envirionment_info}
\alias{get_manifest_envirionment_info}
\title{Get Environment information for manifest}
\usage{
get_manifest_envirionment_info()
}
\value{
nested list of file details, length the same as the input vector.
}
\description{
This function takes no arguments and returns a nested list, suitable for
inclusion in manifest export.
}
