% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_environment.R
\name{get_r_session_info}
\alias{get_r_session_info}
\title{Get session information for manifest}
\usage{
get_r_session_info()
}
\value{
list of session details, including R Version, platform, OS
(\code{running}), locale, timezone, and library paths.
}
\description{
This function takes no arguments and returns a list, suitable for
inclusion in manifest export.
}
