% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_metadata.R
\name{get_single_file_metadata}
\alias{get_single_file_metadata}
\title{Get Metadata for a file}
\usage{
get_single_file_metadata(filepath, summary_info = FALSE)
}
\arguments{
\item{filepath}{vector of filepaths}

\item{summary_info}{logical, whether to include summary information
(nrow/length, colnames/names, class)}
}
\value{
list of file details
}
\description{
This function takes a single filepaths and returns a list of file
details, suitable for inclusion in manifest export.
}
