% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_branch_info}
\alias{git_branch_info}
\title{git_branch_info}
\usage{
git_branch_info(repo)
}
\arguments{
\item{repo}{(Local directory) path to a git repository.}
}
\value{
list of branch information. Keys include:
\itemize{
\item name: name of the active branch.
\item commit: commit hash of the active branch.
\item upstream: name of the upstream branch.
\item remote_url: URL of the remote repository.
\item up_to_date: logical indicating if the active branch is up to date with
remote
\item upstream_commit: commit hash of the upstream branch.
}
}
\description{
Get branch information for a repository.
}
