% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_changed_files}
\alias{git_changed_files}
\title{git_changed_files}
\usage{
git_changed_files(repo)
}
\arguments{
\item{repo}{(Local directory) path to a git repository.}
}
\value{
list of changed files. Keys are file paths, values are status.
}
\description{
Get changed files in a repository.
}
