% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_params.R
\name{inherit_params}
\alias{inherit_params}
\title{inherit_params}
\usage{
inherit_params(params, inheritence_search_paths)
}
\arguments{
\item{params}{list of parameters, which may include an inheritence key
(\code{inherit}).}

\item{inheritence_search_paths}{Paths to search for inherited parameters.}
}
\value{
merged list of parameters, with inherited parameters overlaid onto
\code{params}.
}
\description{
Inherit parameters from a JSON file. This function will search for a file
(in \code{inheritence_search_paths}) with the name of the value of the \code{inherit}
key in the \code{params} list. If found, the parameters in that file will be
overlaid onto the \code{params} list. This process will continue until no
\code{inherit} key is found.
}
