% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{is_git_path}
\alias{is_git_path}
\title{is_git_path}
\usage{
is_git_path(path)
}
\arguments{
\item{path}{(Local directory) path to check.}
}
\value{
logical indicating if the path is in a git repository.
}
\description{
Check if a path is in a git repository.
}
