% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lists.R
\name{merge_lists}
\alias{merge_lists}
\title{Merge two lists}
\usage{
merge_lists(base_list, overlay_list, recursive = TRUE)
}
\arguments{
\item{base_list}{a named list}

\item{overlay_list}{a named list}

\item{recursive}{Should list be merged recurisvely, or only with top level
keys?}
}
\value{
merged list
}
\description{
This function takes two lists (\code{base} and \code{overlay}), and merges them (by
default recursively) into a single list. If a key is present in \code{overlay},
it is inherited from \code{overlay}, but keys missing in \code{overlay} will be
inherited from \code{base}
}
\details{
Code is heavily taken from the \code{merge} function in the \code{config} package.
Notable differences between the two functions are:
\itemize{
\item This does not reorder keys in the lists
}
}
