% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_list_element.R
\name{modify_list_element}
\alias{modify_list_element}
\title{modify_list_element

Apply a function to a one or more elements of a list, given the positions of
the elements.}
\usage{
modify_list_element(x, positions, function_to_apply)
}
\arguments{
\item{x}{List to modify}

\item{positions}{(List or vector) Position of elements to modify, as a
vector of indices}

\item{function_to_apply}{Function to apply to elements}
}
\value{
modified list
}
\description{
modify_list_element

Apply a function to a one or more elements of a list, given the positions of
the elements.
}
