% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta.workflow.utils-package.R
\docType{package}
\name{pacta.workflow.utils-package}
\alias{pacta.workflow.utils}
\alias{pacta.workflow.utils-package}
\title{pacta.workflow.utils: Utility functions for PACTA workflows}
\description{
Provide utility functions to be called across RMI-PACTA's workflows.
}
\author{
\strong{Maintainer}: Alex Axthelm \email{aaxthelm@rmi.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]

Other contributors:
\itemize{
  \item RMI \email{PACTA4investors@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
