% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_params.R
\name{parse_raw_params}
\alias{parse_raw_params}
\title{parse_raw_params}
\usage{
parse_raw_params(
  json,
  inheritence_search_paths = NULL,
  schema_file = NULL,
  raw_schema_file = NULL
)
}
\arguments{
\item{json}{JSON string}

\item{inheritence_search_paths}{value}

\item{schema_file}{value}

\item{raw_schema_file}{value}
}
\value{
returndes
}
\description{
Parse and validate JSON parameters, including optionally validation raw
parameters prior to inheriting.
}
