% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_manifest.R
\name{un_asis}
\alias{un_asis}
\title{un_asis}
\usage{
un_asis(x)
}
\arguments{
\item{x}{an object (with the \code{AsIs} class)}
}
\value{
the same object, without AsIs class
}
\description{
Remove AsIs class from object
}
