% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_manifest.R
\name{export_manifest}
\alias{export_manifest}
\title{Export manifest file with metadata}
\usage{
export_manifest(manifest_path, input_files, output_files)
}
\arguments{
\item{manifest_path}{Path to the manifest file.}

\item{input_files}{List of files that are inputs to the workflow.}

\item{output_files}{List of files that are outputs from the workflow.}
}
\value{
(invisible) JSON string with metadata manifest.
}
\description{
Export manifest file with metadata
}
