% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_metadata.R
\name{get_file_metadata}
\alias{get_file_metadata}
\title{Get Metadata for a vector of filepaths}
\usage{
get_file_metadata(filepaths)
}
\arguments{
\item{filepaths}{vector of filepaths}
}
\value{
nested list of file details, length the same as the input vector.
}
\description{
This function takes a vector of filepaths and returns a list of file
details, suitable for inclusion in manifest export.
}
