% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_info.R
\name{get_package_info}
\alias{get_package_info}
\title{Get package information for active packages}
\usage{
get_package_info(
  packagelist = list(base = utils::sessionInfo()[["basePkgs"]], attached =
    names(utils::sessionInfo()[["otherPkgs"]]), loaded =
    names(utils::sessionInfo()[["loadedOnly"]]))
)
}
\arguments{
\item{packagelist}{vector or list of package names. Best left as default,
which includes the currently loaded and attached namespaces separated into
useful categories.}
}
\value{
nested list of file details, length 3, with top level keys being
\code{base}, \code{attached}, and \code{loaded}. Underneath those keys are lists of package
details, with the package names as keys, and further details as returned by
\code{\link[=get_individual_package_info]{get_individual_package_info()}}.
}
\description{
This function takes a vector or (possibly nested) list of package names and
returns a nested list of package details, suitable for inclusion in manifest
export.
}
\seealso{
\code{\link[=get_individual_package_info]{get_individual_package_info()}}
}
