% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_list_element.R
\name{modify_list_element}
\alias{modify_list_element}
\title{title modify_list_element}
\usage{
modify_list_element(x, positions, function_to_apply)
}
\arguments{
\item{x}{List to modify}

\item{positions}{(List or vector) Position of elements to modify, as a
vector of indices}

\item{function_to_apply}{Function to apply to elements}
}
\value{
modified list
}
\description{
Apply a function to a one or more elements of a list, given the positions of
the elements.
}
\examples{
test_list <- list(a = 1L, b = list(ba = 2L, bb = 2.2), c = "a")
results <- modify_list_element(
  x = test_list,
  position = list(
    c("b", "bb"),
    "a"
  ),
  function_to_apply = I
)
results
}
