% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_list_element.R
\name{modify_single_list_element}
\alias{modify_single_list_element}
\title{modify_single_list_element

Apply a function to a single element of a list, givin a vector of list
indices.}
\usage{
modify_single_list_element(x, position, function_to_apply)
}
\arguments{
\item{x}{List to modify}

\item{position}{Position of element to modify, as a vector of indices}

\item{function_to_apply}{Function to apply to element}
}
\value{
modified list
}
\description{
modify_single_list_element

Apply a function to a single element of a list, givin a vector of list
indices.
}
