% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_params.R
\name{parse_params}
\alias{parse_params}
\title{parse_params}
\usage{
parse_params(
  json,
  inheritence_search_paths = NULL,
  schema_file = NULL,
  force_array = list(c("portfolio", "files"))
)
}
\arguments{
\item{json}{JSON string or file path.}

\item{inheritence_search_paths}{Paths to search for inherited parameters.
See \code{inherit_params}.}

\item{schema_file}{Path to JSON Schema file for validation.}

\item{force_array}{Path in params list to force casting as JSON array.}
}
\value{
Parsed parameters as a standard R list.
}
\description{
Parse parameters from JSON string or file.
}
